Move webdav tests to new class
